Update nth-fibonacci.js
Write a number that takes an integer and returns the nth Fibonacci number. The Fibonacci sequence begins with 0 and 1 as its first and second numbers. Each subsequent nth number is equal to f(n - 1) + f(n - 2). In other words, each nth number is equal to the Fibonacci number before itself added to the Fibonacci number before the number which came before the nth number.